requirements.txt
appdirs==1.4.0
Django==1.11.13
django-cors-headers==2.0.2
djangorestframework==3.8.2
gunicorn==19.6.0
packaging==16.8
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
whitenoise==3.2.3
dj-database-url==0.4.2
drf-nested-routers==0.90.2
shortuuid==0.5.0
django-rest-auth==0.9.3
django-allauth==0.36.0
sendgrid-django==4.2.0
psycopg2-binary==2.7.4
coreapi==2.3.3
django-rest-swagger==2.2.0